requirements.txt
pyspark==3.5.0
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.8
streamlit==1.29.0
pillow==10.1.0
